+++ applications/demo/src/App.tsx
@@ -1,8 +1,8 @@
 // import React from 'react';
 import './App.css';
 import { createRoot } from '@urrri/hi-mobx';
-import { TodoList } from './features/TodoList';
-import { TodoNew } from './features/TodoNew';
+import { TodoList } from './views/TodoList';
+import { TodoNew } from './views/TodoNew';
 import { StoreProvider } from './stores/storeProvider';
 import { RootStore } from './stores/rootStore';
