+++ applications/demo/src/utils/useEnter.ts
@@ -1,4 +1,4 @@
-import { KeyboardEvent } from 'react';
+import type { KeyboardEvent } from 'react';
 
 export const onEnterPress = (cb: any) => (e: KeyboardEvent) => {
   if (e.key === 'Enter') {
